push to test new version of paco

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: jcesar-s <jcesar-s@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 17:44:41 by jcesar-s          #+#    #+#             */
-/*   Updated: 2025/04/10 18:07:58 by jcesar-s         ###   ########.fr       */
+/*   Updated: 2025/04/16 15:32:02 by jcesar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	size_t			i;
-	unsigned long	src_len;
+	size_t	i;
+	size_t	src_len;
 
 	src_len = ft_strlen(src);
 	if (size == 0)
-		return ((size_t)src_len);
+		return (src_len);
 	i = 0;
 	while (src[i] && i < size - 1)
 	{
@@ -27,5 +27,5 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 		i++;
 	}
 	dst[i] = 0;
-	return ((size_t)src_len);
+	return (src_len);
 }
