+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jcesar-s <jcesar-s@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 21:23:05 by jcesar-s          #+#    #+#             */
-/*   Updated: 2025/04/09 15:51:13 by jcesar-s         ###   ########.fr       */
+/*   Updated: 2025/04/16 15:34:33 by jcesar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	unsigned long	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
